ft: split out meerkat helper methods into separate files
