+++ meerkat/Meerkat.Indexing.cs
@@ -1,22 +1,15 @@
 ﻿using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
 using System.Linq;
-using System.Linq.Expressions;
-using System.Threading;
-using System.Threading.Tasks;
 using meerkat.Attributes;
 using meerkat.Constants;
 using meerkat.Extensions;
 using MongoDB.Driver;
-using MongoDB.Driver.Linq;
-using MongoUrlParser;
 
 namespace meerkat;
 
 public static partial class Meerkat
 {
-    private static void HandleUniqueIndexing<TSchema, TId>(Type type, IMongoCollection<TSchema> collection)
+    private static void HandleIndexing<TSchema, TId>(Type type, IMongoCollection<TSchema> collection)
         where TSchema : Schema<TId> where TId : IEquatable<TId>
     {
         var typeName = type.FullName;
@@ -24,6 +17,13 @@ public static partial class Meerkat
         if (SchemasWithCheckedIndices.ContainsKey(typeName))
             return;
 
+        HandleUniqueIndexing(type, collection);
+
+        SchemasWithCheckedIndices[typeName] = true;
+    }
+    
+    private static void HandleUniqueIndexing<TSchema>(Type type, IMongoCollection<TSchema> collection)
+    {
         var properties = type.AttributedWith<UniqueIndexAttribute>();
         var indices = properties
             .Select(x =>
@@ -36,7 +36,5 @@ public static partial class Meerkat
 
         if (indices.Any())
             collection.Indexes.CreateMany(indices);
-
-        SchemasWithCheckedIndices[typeName] = true;
     }
 }
